Extract common `io.spine.server.event.Policy`
Also:
  * Move `Just` under `io.spine.server.event`.
