+++ compiler/src/main/kotlin/io/spine/server/event/Just.kt
@@ -24,30 +24,30 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package io.spine.protodata.plugin
+package io.spine.server.event
 
-import com.google.common.truth.Truth.assertThat
-import io.spine.base.Identifier.newUuid
-import io.spine.protodata.test.ProjectCreated
-import org.junit.jupiter.api.Test
+import io.spine.base.EventMessage
+import io.spine.server.tuple.Tuple
 
-class `Just should` {
+/**
+ * A tuple of one event.
+ *
+ * Used when returning an iterable from a handler method for better readability over `List<E>`.
+ */
+public class Just<E : EventMessage>(event: E) : Tuple(event) {
 
-    @Test
-    fun `store one value`() {
-        val just = Just(validEvent)
-        assertThat(just)
-            .containsExactly(validEvent)
-    }
+    public companion object {
 
-    @Test
-    fun `initialize via static method`() {
-        val just = Just.just(validEvent)
-        assertThat(just)
-            .containsExactly(validEvent)
+        private const val serialVersionUID: Long = 0L
+
+        /**
+         * A factory method for Java.
+         *
+         * Prefer the primary constructor in Kotlin.
+         *
+         * This method is intended to be imported statically.
+         */
+        @JvmStatic
+        public fun <E : EventMessage> just(event: E): Just<E> = Just(event)
     }
 }
-
-private val validEvent = ProjectCreated.newBuilder()
-    .setId(newUuid())
-    .build()
