+++ test-env/src/main/kotlin/io/spine/protodata/test/GreedyPolicy.kt
@@ -28,7 +28,7 @@ package io.spine.protodata.test
 
 import io.spine.core.External
 import io.spine.protodata.TypeEntered
-import io.spine.protodata.plugin.Just
+import io.spine.server.event.Just
 import io.spine.protodata.plugin.Policy
 import io.spine.server.event.React
 import io.spine.server.model.Nothing
